修改用户管理，系统内置用户不允许被删除

+++ src/Company.Project.Domain/Authorization/Users/UserManager.cs
@@ -259,6 +259,11 @@ namespace Company.Project.Authorization.Users
 
             foreach (var user in users)
             {
+                if (user.IsSystem)
+                {
+                    throw new UserFriendlyException("不能删除系统用户!");
+                }
+
                 var identityResult = await this.DeleteAsync(user);
                 if (!identityResult.Succeeded)
                 {
